add options struct to initialization
also slightly changes behavior such that unexported fields no longer
cause the program to panic (we skip looking for them in the environment,
as well).

Options struct allows three new options to be set on Init.

1) AllOptional will force the parse to consider each field optional by
default, rather than erroring when it comes across a non-optional field
that is empty.

2) LeaveNil will make the parse leave pointers in a nil state if no
inner fields or values are found after coming across a nil pointer.
If it does come across any non-empty values, then the behavior is to
create a new pointer and fill in that field. This works for any
arbitrary depth, all parent pointers will be created as needed, no
matter how deeply nested the non-empty value is.

3) AllowUnexported allows for unexported fields to be present inside
of the config. We will not set or look for any unexported fields, they
will simply be skipped over.